+++ test/extensions/common/tap/admin_test.cc
@@ -72,7 +72,7 @@ private:
 
 class BaseAdminHandlerTest : public testing::Test {
 public:
-  void setup(Network::Address::Type socket_type = Network::Address::Type::Ip, uint64_t max_concurrent_streams = 1) {
+  void setup(Network::Address::Type socket_type = Network::Address::Type::Ip, uint32_t max_concurrent_streams = 1) {
     ON_CALL(admin_.socket_, addressType()).WillByDefault(Return(socket_type));
     EXPECT_CALL(admin_, addHandler("/tap", "tap filter control", _, true, true, _))
         .WillOnce(DoAll(SaveArg<2>(&cb_), Return(true)));
